+++ src/main/java/com/kpi/kolesnyk/practicum/service/CodeRunner.java
@@ -1,5 +1,5 @@
 package com.kpi.kolesnyk.practicum.service;
 
 public interface CodeRunner {
-    String estimate(Integer taskId, String code);
+    String estimate(Long taskId, String code);
 }
